requirements.txt
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
click==8.1.7
enum34==1.1.10
gitdb==4.0.11
GitPython==3.1.42
pycparser @ file:///private/tmp/pycparser-20231211-25057-cypg40/pycparser-2.21
pywatchman==2.0.0
PyYAML==6.0.1
setuptools==69.2.0
smmap==5.0.1
-e git+ssh://git@github.com/rimian/tisbutascratch.git@32122742813e09642a9050412f0236044449eea4#egg=tisbutascratch
typer==0.9.0
typing_extensions==4.10.0
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.2_1/libexec/wheel-0.42.0-py3-none-any.whl#sha256=5f1ca525bef7466225922019d72cb06a9e91cd3ff807ddd4be1acb84655e5107
